mfd: arizona: Change fast_start pdata name to better reflect functionality
The bit in the register enables MICBIAS fast startup when clear not when
set. This patch changes the name of this pdata option to soft_start to
better match the functionality. We rename rather than invert the
handling to keep the same default functionality, which is fast start
active.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
(cherry picked from commit f773fc6dca4619bdf8da767eaba101a83b766059)

Change-Id: I8785b3a0c237e5ff9c40ab6e6e29dbe443e8b006
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>