ASoC: wm_adsp: Add support for grouped ALSA binary controls
Currently the ALSA ABI imposes a hard limit of 512 bytes per binary
control.  To support coefficient data blocks of larger sizes we carve up
this space into multiple alsa controls.  All of these controls are
identified by a common prefix and suffix of the form ":<blk id>".

Control groupings can also consist of a single block in which case the
suffix ":0" is used.

Change-Id: Ib8560bbd98425b1732b2ccf372a53102bdeb7d7f
Signed-off-by: Dimitris Papastamos <dp@opensource.wolfsonmicro.com>