ASoC: wm_adsp: Move ADSP2 cores to be powered as a DAPM pre event
PGAs often control the output from a chip and if the DSP is also marked
as a PGA it may be powered up after the output has been enabled. This
patch changes the ADSP2 cores to be powered up during the DAPM pre
event so they are powered up before any outputs.

Move the powering up to a pre-event is also a step along the path of
factoring out the DSP core bring up so it can be run asynchronously to
the rest of the path bring up and it would be where we would wait for
the initial bring completion to finish.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>